Refactors time type to `uint32_t`
Ensures timer countdown time values are handled as unsigned 32-bit integers.
This change prevents potential issues with negative timer values and ensures
compatibility with the underlying hardware/system timer implementation,
which typically uses unsigned integers for representing time intervals.

+++ PAC/common/PAC_info.cpp
@@ -407,7 +407,7 @@ void PAC_info::emulation_off()
     }
 #endif
 //-----------------------------------------------------------------------------
-void PAC_info::set_cycle_time( u_long current_cycle_time )
+void PAC_info::set_cycle_time( uint32_t current_cycle_time )
     {
     cycle_time = current_cycle_time;
     }
